msm: vidc: add 32 bit support in video driver
The commit adds the following
1 Expose 32 bit compatibility ioctl only if the compat config is enabled.
2 Keep DMA mask aligned with api definition.
3 Replace the various division operations with 32 bit compatible modules.

CRs-Fixed: 2546421

+++ drivers/media/platform/msm/vidc/msm_vidc_res_parse.c
@@ -1094,7 +1094,7 @@ static int msm_vidc_setup_context_bank(struct msm_vidc_platform_resources *res,
 		dev->dma_parms =
 			devm_kzalloc(dev, sizeof(*dev->dma_parms), GFP_KERNEL);
 	dma_set_max_seg_size(dev, DMA_BIT_MASK(32));
-	dma_set_seg_boundary(dev, DMA_BIT_MASK(64));
+	dma_set_seg_boundary(dev, (unsigned long)DMA_BIT_MASK(64));
 
 	dprintk(VIDC_DBG, "Attached %s and created mapping\n", dev_name(dev));
 	dprintk(VIDC_DBG,
